requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-adk
google-generativeai
llama-index==0.10.22
llama-index-llms-gemini
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

